# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/registration-requests": {
      "post": {
        "operationId": "RegistrationRequestsController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitRegistrationRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Presentar solicitud de registro",
        "tags": [
          "Usuarios"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/users/registration-requests/{requestId}": {
      "get": {
        "operationId": "RegistrationRequestsController_getById",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar solicitud de registro",
        "tags": [
          "Usuarios"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/users/registration-requests/{requestId}/approve": {
      "post": {
        "operationId": "RegistrationRequestsController_approve",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Aprobar solicitud de registro",
        "tags": [
          "Usuarios"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/users/registration-requests/{requestId}/reject": {
      "post": {
        "operationId": "RegistrationRequestsController_reject",
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Rechazar solicitud de registro",
        "tags": [
          "Usuarios"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/users/{userId}": {
      "get": {
        "operationId": "UsersController_getById",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar usuario",
        "tags": [
          "Usuarios"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/users/{userId}/loan-eligibility": {
      "get": {
        "operationId": "UsersController_getEligibility",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar habilitación para préstamos",
        "tags": [
          "Usuarios"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/users/{userId}/trust": {
      "get": {
        "operationId": "UsersController_getTrust",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar perfil de confianza",
        "tags": [
          "Usuarios"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/users/{userId}/sanctions": {
      "get": {
        "operationId": "UsersController_getSanctions",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar sanciones de un usuario",
        "tags": [
          "Usuarios"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/loans": {
      "post": {
        "operationId": "LoansController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registrar préstamo",
        "tags": [
          "Préstamos"
        ],
        "x-implementation-status": "pending"
      }
    },
...
    "/loans/user/{userId}/history": {
      "get": {
        "operationId": "LoansController_listHistory",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar historial de préstamos de un usuario",
        "tags": [
          "Préstamos"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/loans/{loanId}": {
      "get": {
        "operationId": "LoansController_getById",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Consultar préstamo",
        "tags": [
          "Préstamos"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/loans/{loanId}/return": {
      "post": {
        "operationId": "LoansController_registerReturn",
        "parameters": [
          {
            "name": "loanId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnLoanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registrar devolución",
        "tags": [
          "Préstamos"
        ],
        "x-implementation-status": "pending"
      }
    },
...
    "/violations": {
      "post": {
        "operationId": "ViolationsController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterViolationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Registrar incumplimiento",
        "tags": [
          "Reglas y términos"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/violations/{violationId}/appeals": {
      "post": {
        "operationId": "ViolationsController_appeal",
        "parameters": [
          {
            "name": "violationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAppealDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Presentar apelación de un incumplimiento",
        "tags": [
          "Reglas y términos"
        ],
        "x-implementation-status": "pending"
      }
    },
    "/appeals/{appealId}/resolution": {
      "patch": {
        "operationId": "ViolationsController_resolve",
        "parameters": [
          {
            "name": "appealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAppealDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Resolver apelación",
        "tags": [
          "Reglas y términos"
        ],
        "x-implementation-status": "pending"
      }
    },
...
  },
  "info": {
    "title": "CS Préstamos — API",
    "description": "Contratos HTTP declarados en el backend. El estado de implementación se indica por operación.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SubmitRegistrationRequestDto": {
        "type": "object",
        "properties": {
          "userType": {
            "type": "string",
            "enum": [
              "ESTUDIANTE",
              "DOCENTE",
              "ADMINISTRATIVO"
            ]
          },
          "institutionalId": {
            "type": "string",
            "maxLength": 80
          },
          "evidence": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userType"
        ]
      },
      "CreateLoanDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "copyId": {
            "type": "string",
            "format": "uuid"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "userId",
          "copyId",
          "startsAt"
        ]
      },
      "ReturnLoanDto": {
        "type": "object",
        "properties": {
          "returnedAt": {
            "type": "string",
            "format": "date-time"
          },
          "observation": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": [
          "returnedAt"
        ]
      },
...
      "RegisterViolationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "ruleId": {
            "type": "string",
            "format": "uuid"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "userId",
          "ruleId",
          "occurredAt"
        ]
      },
      "SubmitAppealDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "userId",
          "reason",
          "submittedAt"
        ]
      },
      "ResolveAppealDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACEPTADA",
              "RECHAZADA"
            ]
          },
          "resolution": {
            "type": "string",
            "maxLength": 2000
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time"
          },
          "resolvedByAccountId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "status",
          "resolution",
          "resolvedAt",
          "resolvedByAccountId"
        ]
      },
...
    }
  }
}